Goveralls inadvertently executing init
I think since goveralls was rooted inside the project directory that golang was executing the init function on process execution.  This is bad for this little app as I don't plan on doing testing with Rabbit in the picture.  Let's see if travis and drone can execute now that main is calling initialize now instead of relying on the golang init function feature.